[Refactor] Refactor Google Directions with Strategy Pattern
