+++ mobile/src/services/directions/strategies/DirectionsStrategy.ts
@@ -1,0 +1,18 @@
+import type {
+  DirectionsRequest,
+  DirectionsRoute,
+  DirectionsTravelMode,
+} from '../../../types/Directions';
+
+export type DirectionsStrategyContext = {
+  apiKey: string;
+  fetchImpl?: typeof fetch;
+};
+
+export interface DirectionsStrategy {
+  readonly mode: DirectionsTravelMode;
+  fetchRoute(
+    request: DirectionsRequest,
+    context: DirectionsStrategyContext,
+  ): Promise<DirectionsRoute>;
+}
